02-adk/2-tool-agent/tool_agent/pyproject.toml
[project]
name = "2-tool-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.21.0",
]
